fix:making company context
