+++ next.js/context/company-context.tsx
@@ -1,54 +1,88 @@
-// "use client";
+"use client";
 
-// import { createContext, useEffect, useState } from "react";
+import { createContext, useEffect, useState } from "react";
 
-// type Company = {
-//   id: number;
-//   name: string;
-// };
+type Company = {
+  id: number;
+  name: string;
+};
 
-// type CompanyContext = {
-//   myCompany: Company | null;
-// };
+type Product = {
+  company_name: string;
+  id: number;
+  name: string;
+  model_number: string;
+  price: number;
+  description: string;
+};
 
-// export const CompanyContext = createContext<CompanyContext | null>(null);
+type CompanyContext = {
+  myCompany: Company | null;
+  myProducts: Product[];
+  fetchMyProducts: () => Promise<void>;
+};
 
-// export const CompanyContextProvider = ({
-//   children,
-// }: {
-//   children: React.ReactNode;
-// }) => {
-//   const [myCompany, setMyCompany] = useState<Company | null>(null);
+export const CompanyContext = createContext<CompanyContext | null>(null);
 
-//   const fetchMyCompany = async () => {
-//     console.log("ここはきている");
-//     try {
-//       const res = await fetch("http://localhost:3001/api/company/mycompany", {
-//         method: "GET",
-//         credentials: "include",
-//       });
-//       console.log(res);
-//       if (!res.ok) throw new Error("レスポンスエラーが発生");
-//       const data = await res.json();
-//       console.log(data);
-//       setMyCompany(data);
-//     } catch (err) {
-//       console.error(err);
-//     }
-//   };
+export const CompanyContextProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
+  const [myCompany, setMyCompany] = useState<Company | null>(null);
+  const [myProducts, setMyProducts] = useState<Product[]>([]);
 
-//   const contextValue = {
-//     myCompany,
-//   };
+  const fetchMyCompany = async () => {
+    try {
+      const res = await fetch("http://localhost:3001/api/company/mycompany", {
+        method: "GET",
+        credentials: "include",
+      });
+      if (!res.ok) throw new Error("レスポンスエラーが発生");
+      const data = await res.json();
+      setMyCompany(data);
+    } catch (err) {
+      console.error(err);
+    }
+  };
 
-//   useEffect(() => {
-//     console.log("a");
-//     fetchMyCompany();
-//   }, []);
+  const fetchMyProducts = async () => {
+    try {
+      const res = await fetch(
+        "http://localhost:3001/api/company/myproductlist",
+        {
+          method: "GET",
+          credentials: "include", //cookieデータをつけて送る
+        },
+      );
+      if (!res.ok) {
+        throw new Error("[MyCompanyPage]レスポンスエラー(products)");
+      }
+      const data = await res.json();
+      // console.log("[MyCompanyPage]myCompanyデータ取得成功", data);
+      setMyProducts(data);
+    } catch (err) {
+      console.error("[MycompanyPage]myCompanyデータ取得エラー", err);
+    }
+  };
 
-//   return (
-//     <CompanyContext.Provider value={contextValue}>
-//       {children}
-//     </CompanyContext.Provider>
-//   );
-// };
+  useEffect(() => {
+    fetchMyCompany();
+  }, []);
+
+  useEffect(() => {
+    fetchMyProducts();
+  }, [myCompany]);
+
+  const contextValue = {
+    myCompany,
+    myProducts,
+    fetchMyProducts,
+  };
+
+  return (
+    <CompanyContext.Provider value={contextValue}>
+      {children}
+    </CompanyContext.Provider>
+  );
+};
