+++ next.js/src/app/company/CompanyLayout.tsx
@@ -1,11 +1,14 @@
+import { CompanyContextProvider } from "@/context/company-context";
 import { CompanyNav } from "components/navigation/CompanyNav";
 
 export const CompanyLayout = ({ children }: { children: React.ReactNode }) => {
   return (
     <div>
-      <CompanyNav />
-      <hr />
-      <main>{children}</main>
+      <CompanyContextProvider>
+        <CompanyNav />
+        <hr />
+        <main>{children}</main>
+      </CompanyContextProvider>
     </div>
   );
 };
